Resolve bugged subs when overplaying voice line
When playing a voice line on the same frame as the last voice line is stopped, the subtitles got cleared erroneously one too many times. This commit resolves the bug by NOT connecting the "voice finished" signal to the `_on_voice_end` method all the time, but only connecting it when the voice is being played, then disconnecting it when it gets stopped or the signal fires once.