"Support" soft hyphens
Since Godot 5.3 has no support for soft hyphens in engine, and manually figuring out the text width with potentially up to 5 fonts in the string is nearly impossible, just cheat and pre-calculate where the question text should break.
Currently only used in one question (t007) to hyphenate the name of the Greek dish.